fix(viewer): split useDrawing2DPersistence.ts under the module-size budget
#4170 pushed the hook to 438 lines with no allowlist row and no CI lane
(its base retargets after #4167, so test.yml has never run on it) — the
gate would have failed the moment it landed on main. It had actually
crossed 400 two commits earlier, at 432 lines.

Split along two seams rather than raising a budget:
- drawingMarkupRestorePrecedence.ts: the hashCache plus the
  hasPersistedMarkupEntryFor/onLocalStorageDecidedFor/notifyDecided
  protocol useDrawingMarkupRestoreOnLoad.ts (#4170) depends on.
- drawingMarkupSave.ts: the raw store subscription that persists the
  five markup fields on every change, plus its own #4159 Bug 2/Bug 4
  history.

useDrawing2DPersistence.ts keeps only the restore hook itself and
re-exports the two functions the sibling #4170 hook imports, so no
other call site changes. All 44 tests across the six affected suites
still pass; Section2DPanel.tsx is untouched.
